streamlit/src/chatbot/requirements.txt
google-cloud-geminidataanalytics
ipython
pandas
protobuf
proto-plus
altair
